add the compute amount for the ticket prices
