+++ Payment.java
@@ -2,9 +2,11 @@ public class Payment {
 
     String paymentMethod;
     int amount;
-    
-    public double computeAmount(int price, int numOfTicket){
-        
+
+    public double computeAmount(int numOfTicket) {
+
+        int price = 250;
+
         double amount = price * numOfTicket;
         return amount;
